Stubbing out basic init functionality
Initial take on the required data types and translating the FDTD.py code
to Rust. Nothing is tested yet but this should be a good base to work
from for developing the package. My current thoughts are to try and
allow for 1-4D simulations that output CSV data that can then be read in
and visualised in Python.